.ipynb_checkpoints/requirements-checkpoint.txt
numpy==1.26.3
pyyaml==6.0.2
opencv-python==4.10.0.84
scikit-image
jax==0.6.2
flax==0.10.6
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
matplotlib
wandb